feat(agno-microsoft-teams): extract Teams interface package
