+++ .github/workflows/publish-python.yml
@@ -20,6 +20,7 @@ on:
         type: choice
         options:
           - agno-agent-builder
+          - agno-microsoft-teams
           - payload-documents-worker-builder
       ref:
         description: "Git ref (tag or sha) to publish (defaults to current branch)"
