+++ backend/agno-microsoft-teams/agno_microsoft_teams/__init__.py
@@ -1,0 +1,37 @@
+from agno_microsoft_teams.attachments import (
+    TEAMS_FILE_DOWNLOAD_INFO,
+    download_attachments,
+)
+from agno_microsoft_teams.interface import (
+    TEAMS_AGENT_RUN_TIMEOUT_S,
+    Teams,
+    TeamsInterface,
+    acquire_bot_token,
+    build_msal_client,
+)
+from agno_microsoft_teams.outbound_media import (
+    MAX_INLINE_ATTACHMENT_BYTES,
+    build_attachments,
+)
+from agno_microsoft_teams.verification import (
+    VerifiedClaims,
+    prime_jwks_cache,
+    verify_teams_jwt,
+    verify_teams_jwt_sync,
+)
+
+__all__ = [
+    "MAX_INLINE_ATTACHMENT_BYTES",
+    "TEAMS_AGENT_RUN_TIMEOUT_S",
+    "TEAMS_FILE_DOWNLOAD_INFO",
+    "Teams",
+    "TeamsInterface",
+    "VerifiedClaims",
+    "acquire_bot_token",
+    "build_attachments",
+    "build_msal_client",
+    "download_attachments",
+    "prime_jwks_cache",
+    "verify_teams_jwt",
+    "verify_teams_jwt_sync",
+]
