+++ release-please-config.json
@@ -68,6 +68,12 @@
       "changelog-path": "CHANGELOG.md",
       "release-type": "python"
     },
+    "backend/agno-microsoft-teams": {
+      "package-name": "agno-microsoft-teams",
+      "component": "agno-microsoft-teams",
+      "changelog-path": "CHANGELOG.md",
+      "release-type": "python"
+    },
     "backend/payload-documents-worker-builder": {
       "package-name": "payload-documents-worker-builder",
       "component": "payload-documents-worker-builder",
